increase amount size

+++ Modules/Purchase/Database/Migrations/2021_08_08_021131_create_purchase_payments_table.php
@@ -16,7 +16,7 @@ class CreatePurchasePaymentsTable extends Migration
         Schema::create('purchase_payments', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('purchase_id');
-            $table->integer('amount');
+            $table->bigInteger('amount');
             $table->date('date');
             $table->string('reference');
             $table->string('payment_method');
